Embed logos as base64 and drop binary assets
