+++ iohblade/.streamlit/config.toml
@@ -1,0 +1,2 @@
+[theme]
+primaryColor = "#1E90FF"
