+++ pyproject.toml
@@ -30,6 +30,9 @@ dependencies = [
 Homepage = "https://github.com/XAI-liacs/BLADE"
 Repository = "https://github.com/XAI-liacs/BLADE"
 
+[project.scripts]
+iohblade-webapp = "iohblade.webapp:main"
+
 [dependency-groups]
 dev = [
     "black>=23.11.0,<24",
@@ -74,6 +77,12 @@ autotuning_methodology = { git = "https://github.com/AutoTuningAssociation/autot
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["iohblade"]
+include = [
+  "iohblade/.streamlit/*",
+]
 
 
 [tool.pytest.ini_options]
